MacOS and Docker Desktop compatibility: use `/private/etc/hosts`, parse Docker output, another way to get CPU and memory info (to be improved though)

+++ src/Filesystem/Filesystem.php
@@ -158,9 +158,13 @@ class Filesystem extends \Symfony\Component\Filesystem\Filesystem implements Pro
             $this->mkdir($dir);
         }
 
+        if ($this->isFile($path, true) && !is_writable($path)) {
+           throw new IOException("File is not writeable: $path");
+        }
+
         // Writing empty file returns integer 0
         if (file_put_contents($path, $content, $flags) === false) {
-            throw new IOException("Can't write to file $path!");
+            throw new IOException("Failed to write content to file: $path");
         }
     }
 
@@ -174,6 +178,11 @@ class Filesystem extends \Symfony\Component\Filesystem\Filesystem implements Pro
         $this->firewall($files);
 
         parent::remove($files);
+    }
+
+    public function getHostsFilePath(): string
+    {
+        return PHP_OS_FAMILY === 'Darwin' ? '/private/etc/hosts' : '/etc/hosts';
     }
 
     /**
@@ -210,13 +219,9 @@ class Filesystem extends \Symfony\Component\Filesystem\Filesystem implements Pro
 
         $allowedPaths = [
             sys_get_temp_dir(),
-            '/etc/hosts',
+            $this->getHostsFilePath(),
             $this->env->getProjectsRootDir()
         ];
-
-        if ($this->isMacOs()) {
-            $allowedPaths[] = '/private/etc/hosts';
-        }
 
         // @TODO: Commands need some ACL to define which paths they require to modify or they can ommit?
         // `docker:mysql:reconstruct-db` does not need SSL certificates dir or editing `/etc/hosts` file
@@ -245,13 +250,5 @@ class Filesystem extends \Symfony\Component\Filesystem\Filesystem implements Pro
                 ));
             }
         }
-    }
-
-    /**
-     * @return bool
-     */
-    private function isMacOs(): bool
-    {
-        return PHP_OS_FAMILY === 'Darwin';
     }
 }
